+++ src/lib.rs
@@ -138,7 +138,6 @@ mod macros;
 
 #[cfg(target_arch = "arm")]
 mod lang_items;
-mod l3gd20;
 
 pub mod delay;
 pub mod examples;
@@ -146,6 +145,7 @@ pub mod exception;
 pub mod fpu;
 pub mod interrupt;
 pub mod itm;
+pub mod l3gd20;
 pub mod led;
 pub mod lsm303dlhc;
 pub mod peripheral;